Cryptomind.OCR/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0